+++ adapters/InMobi/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## InMobi iOS Mediation Adapter Changelog
 
+#### Next version
+- Replaced the deprecated method GADMobileAds.sharedInstance.sdkVersion with GADMobileAds.sharedInstance.versionNumber.
+
 #### [Version 10.6.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/inmobi/InMobiAdapter-10.6.0.0.zip)
 - Verified compatibility with InMobi SDK 10.6.0.
 - Added watermark support for bidding banner, interstitial and rewarded ad formats.